requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
transformers>=4.30.0
faiss-cpu>=1.7.0
onnxruntime>=1.15.0
easyocr>=1.7.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.7.0
pyyaml>=6.0
streamlit>=1.28.0
pillow>=9.0.0
scikit-learn>=1.3.0
yolov7>=0.1.0
sentence-transformers>=2.2.0
clip-by-openai>=1.0
